Added some comments in lib/madmask.js and fixed bugs in android-sdk.spec

+++ HappyFace-MadMask/SPECS/android-sdk.spec
@@ -8,6 +8,7 @@ URL: https://dl.google.com
 BuildRoot: %{_tmppath}/%{name}-%{version}-root
 Requires: java-1.7.0-openjdk-devel
 Requires: ant-apache-regexp
+Requires: glibc
 
 
 
@@ -17,7 +18,8 @@ Requires: ant-apache-regexp
 # Preamble
 #
 # Macro definitions
-%define _prefix         /usr/local
+%define _prefix         /usr/local/android-tools
+%define _bin_dir        /usr/bin
 %define _profile_dir    /etc/profile.d
 
 %description
@@ -36,6 +38,7 @@ cd ..
 
 # make directories
 ! [ -d $RPM_BUILD_ROOT/%{_prefix} ] && mkdir -vp $RPM_BUILD_ROOT/%{_prefix}
+! [ -d $RPM_BUILD_ROOT/%{_bin_dir} ] && mkdir -vp $RPM_BUILD_ROOT/%{_bin_dir}
 ! [ -d $RPM_BUILD_ROOT/%{_profile_dir} ] && mkdir -vp $RPM_BUILD_ROOT/%{_profile_dir}
 
 
@@ -47,20 +50,39 @@ export ANT_HOME=%{_prefix}/apache-ant
 export PATH=\$ANDROID_HOME/platforms:\$ANDROID_HOME/platform-tools:\$ANDROID_HOME/tools:\$ANT_HOME:$ANT_HOME/bin:\$PATH
 " > $RPM_BUILD_ROOT/%{_profile_dir}/android-sdk.sh
 
-## Ant
-ANT="http://apache.lauf-forum.at//ant/binaries/apache-ant-1.9.10-bin.zip"
-if [ ! -e /tmp/$(basename $ANT) ]; then
-    wget "$ANT" -O /tmp/$(basename $ANT) || rm -v /tmp/$(basename $ANT)
+
+## A script installing the latest SDK packages
+echo "#!/bin/bash
+source %{_profile_dir}/android-sdk.sh
+(while sleep 3; do echo y; done) | android update sdk --no-ui --all
+" > $RPM_BUILD_ROOT/%{_bin_dir}/update-android-sdk
+chmod 755 $RPM_BUILD_ROOT/%{_bin_dir}/update-android-sdk
+
+
+%post
+
+## Installing Android SDK Tools
+if [ ! -e %{_prefix}/android ]; then
+    ## Making symlinks
+    ln -vs apache-ant-1.9.10 %{_prefix}/apache-ant
+    ln -vs adt-bundle-linux-x86_64-20140702 %{_prefix}/android
+
+    ## Ant
+    ANT="http://apache.lauf-forum.at//ant/binaries/apache-ant-1.9.10-bin.zip"
+    if [ ! -e /tmp/$(basename $ANT) ]; then
+	wget "$ANT" -O /tmp/$(basename $ANT) || rm -v /tmp/$(basename $ANT)
+    fi
+    
+    ## Android SDK
+    ADT="https://dl.google.com/android/adt/adt-bundle-linux-x86_64-20140702.zip"
+    if [ ! -e /tmp/$(basename $ADT) ]; then
+	wget "$ADT" -O /tmp/$(basename $ADT) || rm -v /tmp/$(basename $ADT)
+    fi
+    
+    unzip /tmp/$(basename $ANT) -d %{_prefix} && 
+    [ ! -e %{_prefix}/android ] && unzip /tmp/$(basename $ADT) -d %{_prefix}
 fi
 
-## Android SDK
-ADT="https://dl.google.com/android/adt/adt-bundle-linux-x86_64-20140702.zip"
-if [ ! -e /tmp/$(basename $ADT) ]; then
-    wget "$ADT" -O /tmp/$(basename $ADT) || rm -v /tmp/$(basename $ADT)
-fi
-
-[ ! -e %{_prefix}/apache-ant ] && unzip /tmp/$(basename $ANT) -d $RPM_BUILD_ROOT/%{_prefix} && ln -s %{_prefix}/apache-ant-1.9.10 $RPM_BUILD_ROOT/%{_prefix}/apache-ant
-[ ! -e %{_prefix}/android ] && unzip /tmp/$(basename $ADT) -d $RPM_BUILD_ROOT/%{_prefix} && ln -s %{_prefix}/adt-bundle-linux-x86_64-20140702 $RPM_BUILD_ROOT/%{_prefix}/android
 
 
 
@@ -70,7 +92,8 @@ fi
 
 %files
 %defattr(-,root,root)
-%{_prefix}/*
+%{_prefix}
+%{_bin_dir}/*
 %{_profile_dir}/*
